requirements.txt
SQLAlchemy==2.0.29
aiohttp==3.9.3
Wikipedia-API==0.6.0
requests==2.31.0
wikipedia==1.4.0
python-telegram-bot==21.0.1
translators==5.9.0
matplotlib==3.8.4
prettytable==3.10.0